bigquery-proxy/_src/requirements.txt
requests
google-cloud-bigquery
bson
